backend/requirements.txt
fastapi[standard]
python-jose[cryptography]
passlib[argon2]
sqlmodel
alembic
python-dotenv
aiosqlite
greenlet
numpy
pandas
scikit-learn
joblib
sentence-transformers